requirements.txt
tkinter
requests
datetime
pyperclip

pysocks

urllib3
